feat: hugely improve scheduler utilization w/ `ets`
i knew i had a scheduling bottleneck due to the number of messages each
scheduled loader was receiving after i changed the client API for
declaring a load profile so that the user did not need to tell the
library how to batch things. but `:observer` showed me **just how bad
that bottleneck was**.

so i came up with a few alternate designs and landed on using `ets`,
allowing each `Task` process to write to the table through a client API,
to minimize the amount of messages that the scheduled loader `GenServer`
has to handle. this also enables:
  - a log of scheduled loader executions, for as long as the table
    survives (before this information was lost as soon as the process
    terminated)
  - a mechanism by which a scheduled loader could be restarted _nearly_
    where it left off (depends on if the tasks have side-effects) in the
    case of transient errors, b/c a new loader process could be seeded
    with the state saved in `ets`